requirements.txt
duckduckgo_search==8.0.0
httpx==0.28.1
markdownify==1.1.0
mistralai==1.6.0
pandas==2.0.3
python-dotenv==1.1.0
smolagents==1.13.0
transformers[torch]
bitsandbytes
